move main to cmd package and have rpc package as an actual server
