+++ rpc/io.go
@@ -1,4 +1,4 @@
-package main
+package rpc
 
 import (
 	"bytes"
