+++ rpc/middleware.go
@@ -1,4 +1,4 @@
-package main
+package rpc
 
 import (
 	"errors"
