+++ rpc/utils_test.go
@@ -1,4 +1,4 @@
-package main
+package rpc
 
 import (
 	"net/http"
